+++ app/utils/yaml_editor.py
@@ -82,3 +82,4 @@ class YAMLEditor:
         
         return content, False
 
+